change hand.js to hand.ts
