+++ src/test/js/jan/hand-test.js
@@ -16,7 +16,7 @@ limitations under the License.
 import assert from 'assert';
 
 import {CALL_TYPE} from '../../../main/js/jan/call-type';
-import Hand from '../../../main/js/jan/hand';
+import {Hand} from '../../../main/js/jan/hand';
 import {Janpai} from '../../../main/js/jan/janpai';
 import {JANPAI_ID} from '../../../main/js/jan/janpai-id';
 import {Mentsu} from '../../../main/js/jan/mentsu';
