+++ src/test/js/jan/jan-util-test.js
@@ -15,7 +15,7 @@ limitations under the License.
  */
 import assert from 'assert';
 
-import Hand from '../../../main/js/jan/hand';
+import {Hand} from '../../../main/js/jan/hand';
 import {Janpai} from '../../../main/js/jan/janpai';
 import {JANPAI_ID} from '../../../main/js/jan/janpai-id';
 import JanUtil from '../../../main/js/jan/jan-util';
